fix: durationMs returns integer for Tempo TraceQL search (issue #782)
The TraceQL search path used toFloat64() for computing _duration_ms,
producing fractional JSON values (e.g. 1.637642) that Grafana's Tempo
datasource cannot unmarshal into uint32.

- traces_data.go: replace toFloat64(...) / 1000000 with intDiv(..., 1000000)
- responses.go: change TraceInfo.DurationMs float64 -> int64
- reqest_processor.go: change traceDurationMs float64 -> int64

+++ reader/traceql/traceql_transpiler/reqest_processor.go
@@ -45,7 +45,7 @@ func (t TraceQLRequestProcessor) Process(ctx *shared.PlannerContext) (chan []mod
 				durationsNs       []int64
 				timestampsNs      []int64
 				startTimeUnixNano int64
-				traceDurationMs   float64
+				traceDurationMs   int64
 				rootServiceName   string
 				rootTraceName     string
 			)
